Add permissions for ratifications

+++ ferry/court/api/views/ratifications.py
@@ -7,7 +7,7 @@ from django.http import HttpRequest
 from django.shortcuts import get_object_or_404
 from ninja import Router, errors
 
-from ferry.core.exceptions import ConflictError, InternalServerError
+from ferry.core.exceptions import ConflictError, ForbiddenError, InternalServerError
 from ferry.core.schema import ConfirmationDetail, ErrorDetail
 from ferry.court.api.schema import RatificationCreate, RatificationDetail
 from ferry.court.models import Accusation, Consequence, Person, Ratification
@@ -27,7 +27,12 @@ router = Router(tags=["Ratifications"])
 )
 def ratification_detail(request: HttpRequest, accusation_id: UUID) -> Ratification:
     assert request.user.is_authenticated
-    return get_object_or_404(Ratification, accusation_id=accusation_id)
+    ratification = get_object_or_404(Ratification, accusation_id=accusation_id)
+
+    if not request.user.has_perm("court.view_ratification", ratification):
+        raise ForbiddenError()
+
+    return ratification
 
 
 @router.post(
@@ -45,6 +50,9 @@ def ratification_create(request: HttpRequest, accusation_id: UUID, payload: Rati
     assert request.user.is_authenticated
     accusation = get_object_or_404(Accusation, id=accusation_id)
 
+    if not request.user.has_perm("court.create_ratification"):
+        raise ForbiddenError()
+
     if Ratification.objects.filter(accusation=accusation).exists():
         raise ConflictError("Ratification already exists")
 
@@ -54,6 +62,9 @@ def ratification_create(request: HttpRequest, accusation_id: UUID, payload: Rati
         raise errors.ValidationError(
             [{"loc": "created_by", "detail": f"Unable to find person with ID {payload.created_by}"}]
         ) from None
+
+    if not request.user.has_perm("court.act_on_behalf_of_person", creator):
+        raise ForbiddenError("You cannot act on behalf of other people.")
 
     try:
         consequence = random.choice(Consequence.objects.filter(is_enabled=True).all())  # noqa: S311
@@ -87,6 +98,9 @@ def ratification_create(request: HttpRequest, accusation_id: UUID, payload: Rati
 def ratification_delete(request: HttpRequest, accusation_id: UUID) -> ConfirmationDetail:
     assert request.user.is_authenticated
     ratification = get_object_or_404(Ratification, accusation_id=accusation_id)
+
+    if not request.user.has_perm("court.delete_ratification", ratification):
+        raise ForbiddenError()
 
     ratification.delete()
     return ConfirmationDetail()
